project/requirements.txt
django
django-rest-framework
django-cors-headers
psycopg2-binary
setuptools
coreapi
pillow